+++ src/StateSmith/Input/antlr4/ErrorListener.cs
@@ -4,7 +4,7 @@ using System.IO;
 
 namespace StateSmith.Input.antlr4
 {
-    public class ErrorListener : IAntlrErrorListener<IToken>
+    public class ErrorListener : IAntlrErrorListener<IToken> /* for parser errors */, IAntlrErrorListener<int> /* for lexer errors */
     {
         public List<Error> errors = new List<Error>();
 
@@ -12,6 +12,18 @@ namespace StateSmith.Input.antlr4
         {
             errors.Add(new Error() {
                 offendingSymbol = offendingSymbol,
+                line = line,
+                column = charPositionInLine,
+                message = msg,
+                exception = e
+            });
+        }
+
+        public void SyntaxError(TextWriter output, IRecognizer recognizer, int offendingChar, int line, int charPositionInLine, string msg, RecognitionException e)
+        {
+            errors.Add(new Error()
+            {
+                offendingChar = offendingChar,
                 line = line,
                 column = charPositionInLine,
                 message = msg,