+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - detect yEd hidden edges https://github.com/StateSmith/StateSmith/issues/29
 - Better error reporting regarding diagram edge IDs.
+- Improve PlantUML parsing https://github.com/StateSmith/StateSmith/issues/21
+  - `notes`, and improve parsing of malformed `skinparam` blocks.
+- Prevent antlr4 error output to console.
+
+
+### Fixed
+- Now throws useful lexer stage exceptions instead of printing them and not recognizing the failure. https://github.com/StateSmith/StateSmith/issues/31 
 
 ## [0.5.6-alpha]
 ### Added